Updated DataViewKitchenSink.tsx to work with storysource addon
